requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor-5==0.2.2
django-filter==22.1
django_debug_toolbar==3.8.1
Pillow==9.4.0
sqlparse==0.4.3
